+++ DifferentiationInterface/ext/DifferentiationInterfaceSymbolicsExt/onearg.jl
@@ -15,12 +15,13 @@ function DI.prepare_pushforward_nokwarg(
     dx_var = variablize(dx, :dx)
     t_var = variable(:t)
     context_vars = variablize(contexts)
+    context_args = argumentize(context_vars, contexts)
     step_der_var = derivative(f(x_var + t_var * dx_var, context_vars...), t_var)
     pf_var = substitute(step_der_var, Dict(t_var => zero(eltype(x))))
 
     erase_cache_vars!(context_vars, contexts)
     res = build_function(
-        pf_var, x_var, dx_var, context_vars...; expression = Val(false), cse = true
+        pf_var, x_var, dx_var, context_args...; expression = Val(false), cse = true
     )
     (pf_exe, pf_exe!) = if res isa Tuple
         res
@@ -103,10 +104,11 @@ function DI.prepare_derivative_nokwarg(
     _sig = DI.signature(f, backend, x, contexts...; strict)
     x_var = variablize(x, :x)
     context_vars = variablize(contexts)
+    context_args = argumentize(context_vars, contexts)
     der_var = derivative(f(x_var, context_vars...), x_var)
 
     erase_cache_vars!(context_vars, contexts)
-    res = build_function(der_var, x_var, context_vars...; expression = Val(false), cse = true)
+    res = build_function(der_var, x_var, context_args...; expression = Val(false), cse = true)
     (der_exe, der_exe!) = if res isa Tuple
         res
     elseif res isa RuntimeGeneratedFunction
@@ -178,12 +180,13 @@ function DI.prepare_gradient_nokwarg(
     _sig = DI.signature(f, backend, x, contexts...; strict)
     x_var = variablize(x, :x)
     context_vars = variablize(contexts)
+    context_args = argumentize(context_vars, contexts)
     # Symbolic.gradient only accepts vectors
     grad_var = gradient(f(x_var, context_vars...), vec(x_var))
 
     erase_cache_vars!(context_vars, contexts)
     res = build_function(
-        grad_var, vec(x_var), context_vars...; expression = Val(false), cse = true
+        grad_var, vec(x_var), context_args...; expression = Val(false), cse = true
     )
     (grad_exe, grad_exe!) = res
     return SymbolicsOneArgGradientPrep(_sig, grad_exe, grad_exe!)
@@ -256,6 +259,7 @@ function DI.prepare_jacobian_nokwarg(
     _sig = DI.signature(f, backend, x, contexts...; strict)
     x_var = variablize(x, :x)
     context_vars = variablize(contexts)
+    context_args = argumentize(context_vars, contexts)
     if backend isa AutoSparse
         jac_var = sparsejacobian(vec(f(x_var, context_vars...)), vec(x_var))
         sparsity = DI.get_pattern(jac_var)
@@ -265,7 +269,7 @@ function DI.prepare_jacobian_nokwarg(
     end
 
     erase_cache_vars!(context_vars, contexts)
-    res = build_function(jac_var, x_var, context_vars...; expression = Val(false), cse = true)
+    res = build_function(jac_var, x_var, context_args...; expression = Val(false), cse = true)
     (jac_exe, jac_exe!) = res
     return SymbolicsOneArgJacobianPrep(_sig, sparsity, jac_exe, jac_exe!)
 end
@@ -338,6 +342,7 @@ function DI.prepare_hessian_nokwarg(
     _sig = DI.signature(f, backend, x, contexts...; strict)
     x_var = variablize(x, :x)
     context_vars = variablize(contexts)
+    context_args = argumentize(context_vars, contexts)
     # Symbolic.hessian only accepts vectors
     if backend isa AutoSparse
         hess_var = sparsehessian(f(x_var, context_vars...), vec(x_var))
@@ -349,7 +354,7 @@ function DI.prepare_hessian_nokwarg(
 
     erase_cache_vars!(context_vars, contexts)
     res = build_function(
-        hess_var, vec(x_var), context_vars...; expression = Val(false), cse = true
+        hess_var, vec(x_var), context_args...; expression = Val(false), cse = true
     )
     (hess_exe, hess_exe!) = res
 
@@ -432,6 +437,7 @@ function DI.prepare_hvp_nokwarg(
     x_var = variablize(x, :x)
     dx_var = variablize(dx, :dx)
     context_vars = variablize(contexts)
+    context_args = argumentize(context_vars, contexts)
     # Symbolic.hessian only accepts vectors
     hess_var = hessian(f(x_var, context_vars...), vec(x_var))
     hvp_vec_var = hess_var * vec(dx_var)
@@ -441,7 +447,7 @@ function DI.prepare_hvp_nokwarg(
         hvp_vec_var,
         vec(x_var),
         vec(dx_var),
-        context_vars...;
+        context_args...;
         expression = Val(false),
         cse = true,
     )
@@ -528,11 +534,12 @@ function DI.prepare_second_derivative_nokwarg(
     _sig = DI.signature(f, backend, x, contexts...; strict)
     x_var = variablize(x, :x)
     context_vars = variablize(contexts)
+    context_args = argumentize(context_vars, contexts)
     der_var = derivative(f(x_var, context_vars...), x_var)
     der2_var = derivative(der_var, x_var)
 
     erase_cache_vars!(context_vars, contexts)
-    res = build_function(der2_var, x_var, context_vars...; expression = Val(false), cse = true)
+    res = build_function(der2_var, x_var, context_args...; expression = Val(false), cse = true)
     (der2_exe, der2_exe!) = if res isa Tuple
         res
     elseif res isa RuntimeGeneratedFunction